server/api/requirements.txt
# FastAPI и веб
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# База данных
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# InfluxDB
influxdb-client>=1.47.0

# Валидация и сериализация
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0

# Утилиты
python-dotenv>=1.0.1
httpx>=0.28.0

# Аутентификация
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
